Split sourcemap merging logic into its own file.
