+++ packages/babel-core/src/transformation/file/generate.js
@@ -2,10 +2,10 @@
 
 import type { PluginPasses } from "../../config";
 import convertSourceMap, { type SourceMap } from "convert-source-map";
-import sourceMap from "source-map";
 import generate from "@babel/generator";
 
 import type File from "./file";
+import mergeSourceMap from "./merge-map";
 
 export default function generateCode(
   pluginPasses: PluginPasses,
@@ -71,47 +71,4 @@ export default function generateCode(
   }
 
   return { outputCode, outputMap };
-}
-
-function mergeSourceMap(inputMap: SourceMap, map: SourceMap): SourceMap {
-  const inputMapConsumer = new sourceMap.SourceMapConsumer(inputMap);
-  const outputMapConsumer = new sourceMap.SourceMapConsumer(map);
-
-  const mergedGenerator = new sourceMap.SourceMapGenerator({
-    file: inputMapConsumer.file,
-    sourceRoot: inputMapConsumer.sourceRoot,
-  });
-
-  // This assumes the output map always has a single source, since Babel always compiles a
-  // single source file to a single output file.
-  const source = outputMapConsumer.sources[0];
-
-  inputMapConsumer.eachMapping(function(mapping) {
-    const generatedPosition = outputMapConsumer.generatedPositionFor({
-      line: mapping.generatedLine,
-      column: mapping.generatedColumn,
-      source: source,
-    });
-    if (generatedPosition.column != null) {
-      mergedGenerator.addMapping({
-        source: mapping.source,
-
-        original:
-          mapping.source == null
-            ? null
-            : {
-                line: mapping.originalLine,
-                column: mapping.originalColumn,
-              },
-
-        generated: generatedPosition,
-
-        name: mapping.name,
-      });
-    }
-  });
-
-  const mergedMap = mergedGenerator.toJSON();
-  inputMap.mappings = mergedMap.mappings;
-  return inputMap;
 }
